refactor appearance classes
