+++ connect/src/main/java/com/stripe/android/connect/EmbeddedComponentManager.kt
@@ -2,8 +2,11 @@ package com.stripe.android.connect
 
 import android.content.Context
 import android.os.Parcelable
-import androidx.annotation.ColorInt
 import androidx.annotation.RestrictTo
+import com.stripe.android.connect.appearance.Appearance
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.parcelize.Parcelize
 import kotlin.coroutines.resume
@@ -13,7 +16,11 @@ import kotlin.coroutines.resume
 class EmbeddedComponentManager(
     private val configuration: Configuration,
     private val fetchClientSecretCallback: FetchClientSecretCallback,
+    appearance: Appearance,
 ) {
+
+    private val _appearanceFlow = MutableStateFlow(appearance)
+    internal val appearanceFlow: StateFlow<Appearance> get() = _appearanceFlow.asStateFlow()
 
     /**
      * Create a new [PayoutsView] for inclusion in the view hierarchy.
@@ -49,8 +56,8 @@ class EmbeddedComponentManager(
 
     @PrivateBetaConnectSDK
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    fun update(@Suppress("UNUSED_PARAMETER") appearance: Appearance) {
-        throw NotImplementedError("Appearance update functionality is not yet implemented")
+    fun update(appearance: Appearance) {
+        _appearanceFlow.value = appearance
     }
 
     @PrivateBetaConnectSDK
@@ -67,111 +74,4 @@ class EmbeddedComponentManager(
     data class Configuration(
         val publishableKey: String,
     ) : Parcelable
-
-    // Appearance classes
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class Appearance(
-        val colorsLight: Colors = Colors(),
-        val colorsDark: Colors? = null,
-        val cornerRadius: CornerRadius = CornerRadius(),
-        val typography: Typography = Typography(),
-        val buttonPrimaryLight: Button = Button(),
-        val buttonPrimaryDark: Button? = null,
-        val buttonSecondaryLight: Button = Button(),
-        val buttonSecondaryDark: Button? = null,
-        val badgeNeutralLight: Badge = Badge(),
-        val badgeNeutralDark: Badge? = null,
-        val badgeSuccessLight: Badge = Badge(),
-        val badgeSuccessDark: Badge? = null,
-        val badgeWarningLight: Badge = Badge(),
-        val badgeWarningDark: Badge? = null,
-        val badgeDangerLight: Badge = Badge(),
-        val badgeDangerDark: Badge? = null
-    ) : Parcelable
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class Colors(
-        @ColorInt val primary: Int? = null,
-        @ColorInt val background: Int? = null,
-        @ColorInt val text: Int? = null,
-        @ColorInt val secondaryText: Int? = null,
-        @ColorInt val danger: Int? = null,
-        @ColorInt val border: Int? = null,
-        @ColorInt val actionPrimaryText: Int? = null,
-        @ColorInt val actionSecondaryText: Int? = null,
-        @ColorInt val offsetBackground: Int? = null,
-        @ColorInt val formBackground: Int? = null,
-        @ColorInt val formHighlightBorder: Int? = null,
-        @ColorInt val formAccent: Int? = null
-    ) : Parcelable
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class Button(
-        @ColorInt val colorBackground: Int? = null,
-        @ColorInt val colorBorder: Int? = null,
-        @ColorInt val colorText: Int? = null
-    ) : Parcelable
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class Badge(
-        @ColorInt val colorBackground: Int? = null,
-        @ColorInt val colorText: Int? = null,
-        @ColorInt val colorBorder: Int? = null
-    ) : Parcelable
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class CornerRadius(
-        val base: Float? = null,
-        val button: Float? = null,
-        val badge: Float? = null,
-        val overlay: Float? = null,
-        val form: Float? = null
-    ) : Parcelable
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    @Parcelize
-    data class Typography(
-        val fontFamily: String? = null,
-        val fontSizeBase: Float? = null,
-        val headingXl: Style? = null,
-        val headingLg: Style? = null,
-        val headingMd: Style? = null,
-        val headingSm: Style? = null,
-        val headingXs: Style? = null,
-        val bodyMd: Style? = null,
-        val bodySm: Style? = null,
-        val labelMd: Style? = null,
-        val labelSm: Style? = null
-    ) : Parcelable {
-
-        @PrivateBetaConnectSDK
-        @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-        @Parcelize
-        data class Style(
-            val fontSize: Float? = null,
-            val fontWeight: Int? = null,
-            val textTransform: TextTransform = TextTransform.None
-        ) : Parcelable
-    }
-
-    @PrivateBetaConnectSDK
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    enum class TextTransform {
-        None,
-        Uppercase,
-        Lowercase,
-        Capitalize
-    }
 }
